refactor: #194 팀 유저 관리 서비스로 리팩토링
