+++ src/main/java/com/moyorak/api/team/service/TeamUserService.java
@@ -5,15 +5,11 @@ import com.moyorak.api.team.domain.TeamRole;
 import com.moyorak.api.team.domain.TeamUser;
 import com.moyorak.api.team.domain.TeamUserNotFoundException;
 import com.moyorak.api.team.domain.TeamUserStatus;
-import com.moyorak.api.team.dto.TeamUserListRequest;
-import com.moyorak.api.team.dto.TeamUserResponse;
 import com.moyorak.api.team.repository.TeamRepository;
 import com.moyorak.api.team.repository.TeamUserRepository;
 import com.moyorak.config.exception.BusinessException;
-import com.moyorak.global.domain.ListResponse;
 import java.util.Optional;
 import lombok.RequiredArgsConstructor;
-import org.springframework.data.domain.Page;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -71,29 +67,6 @@ public class TeamUserService {
         teamUser.withdraw();
     }
 
-    @Transactional(readOnly = true)
-    public ListResponse<TeamUserResponse> getTeamUsers(
-            final Long userId, final Long teamId, final TeamUserListRequest request) {
-        final TeamUser teamUser =
-                teamUserRepository
-                        .findByUserIdAndTeamIdAndUse(userId, teamId, true)
-                        .orElseThrow(TeamUserNotFoundException::new);
-
-        if (teamUser.isNotApproved()) {
-            throw new TeamUserNotFoundException();
-        }
-
-        if (!teamUser.isTeamAdmin()) {
-            throw new BusinessException("팀 관리자만 조회할 수 있습니다.");
-        }
-
-        final Page<TeamUserResponse> teamUsers =
-                teamUserRepository.findByConditions(
-                        teamId, request.getStatus(), true, request.toRecentPageable());
-
-        return ListResponse.from(teamUsers);
-    }
-
     @Transactional
     public void requestJoin(final Long userId, final Team team) {
         final Optional<TeamUser> isWithdrawnTeamUser =
@@ -105,26 +78,5 @@ public class TeamUserService {
             teamUserRepository.save(
                     TeamUser.create(team, userId, TeamRole.TEAM_MEMBER, TeamUserStatus.PENDING));
         }
-    }
-
-    @Transactional
-    public void approveRequestJoin(final Long userId, final Long teamId, final Long teamMemberId) {
-        final TeamUser teamAdminUser =
-                teamUserRepository
-                        .findByUserIdAndTeamIdAndUse(userId, teamId, true)
-                        .orElseThrow(TeamUserNotFoundException::new);
-        if (!teamAdminUser.isTeamAdmin()) {
-            throw new BusinessException("승인을 하는 주체가, 팀 관리자가 아닙니다.");
-        }
-
-        final TeamUser teamUser =
-                teamUserRepository
-                        .findByIdAndUseAndStatus(teamMemberId, true, TeamUserStatus.PENDING)
-                        .orElseThrow(TeamUserNotFoundException::new);
-        if (!teamUser.isTeam(teamId)) {
-            throw new BusinessException("해당 팀의 팀원이 아닙니다.");
-        }
-
-        teamUser.changeStatus(TeamUserStatus.APPROVED);
     }
 }
